Add systemless preference value initialization
